docs: auth dto 코드 분리
