+++ src/auth/dto/auth-credential.dto copy.ts
@@ -2,7 +2,6 @@ import {
   IsEmail,
   IsNotEmpty,
   IsString,
-  Matches,
   MaxLength,
   MinLength,
 } from 'class-validator';
@@ -29,40 +28,6 @@ export class AuthCredentialsDto {
   @MinLength(3)
   @MaxLength(20)
   passwordCheck: string;
-
-  @IsString()
-  profileUrl: string;
-}
-
-export class AuthSignInDto {
-  @IsEmail()
-  email: string;
-
-  @IsNotEmpty()
-  @IsString()
-  @MinLength(3)
-  @MaxLength(20)
-  password: string;
-}
-
-export class UserKakaoDto {
-  @IsString()
-  @IsNotEmpty()
-  name: string;
-
-  @IsEmail()
-  email: string | null;
-
-  @IsString()
-  @IsNotEmpty()
-  kakaoId: string;
-
-  @IsString()
-  @IsNotEmpty()
-  kakaoAccessToken: string;
-
-  @IsString()
-  provider: string;
 
   @IsString()
   profileUrl: string;
